feat(tests): Add unit tests for the API
The current `test_read_price_data_zone` is not passing on the current
main branch. I believe this being due to changes not having been pushed
as the SQL queries assume that the database is using the old schema.

The current test cases are made with the assumption of how I imagine the
API endpoint would work and unless there are objections I think the API
should strive towards passing these tests.